% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{sparkhaven}
\alias{sparkhaven}
\alias{sparkhaven-package}
\title{Read SAS, SPSS, & STATA data files into Spark DataFrames}
\description{
'sparkhaven' is an extension for sparklyr to read SAS, SPSS, & STATA data files into Spark DataFrames. It uses different Spark packages to load such datasets in parallel in Spark.
}

